+++ CHANGELOG.md
@@ -7,6 +7,15 @@ This project *surprisingly* adheres to [Semantic Versioning](http://semver.org).
 
 No notable changes.
 
+## [1.1.0] - 2017-02-13
+
+### Added
+
+* `SemanticVersionInterface->isMajorRelease()`
+* `SemanticVersionInterface->isMinorRelease()`
+* `SemanticVersionInterface->isPatchRelease()`
+* `SemanticVersionInterface->isPreRelease()`
+
 ## [1.0.0] - 2016-07-24
 
 ### Removed
@@ -35,5 +44,6 @@ No notable changes.
 * Renamed `NoSemanticVersionException` to `InvalidVersionException`
 
 [Unreleased]: https://github.com/Rayne/semantic-versioning.php/compare/1.0.0...HEAD
+[1.1.0]: https://github.com/Rayne/semantic-versioning.php/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/Rayne/semantic-versioning.php/compare/1.0.0-rc.3...1.0.0
 [1.0.0-rc.3]: https://github.com/Rayne/semantic-versioning.php/compare/1.0.0-rc.2...1.0.0-rc.3